Refactor: File renamed, identation fixes and URL validation
It was added one more step on handling the user's input. Now, possible space characters of the raw URL edges' are removed before checking if it's valid with a regular expression. That RegEx doesn't need communication protocol as mandatory for classifying an input as valid (e.g., 'google.com' is valid).  Finally, some explanatory comments were added to eliminate possible misunderstandings of the source code.